requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
black==24.4.2
bleach==6.1.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==4.2.14
django-debug-toolbar==4.4.6
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
exceptiongroup==1.2.2
filelock==3.15.4
identify==2.6.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
mysqlclient==2.2.4
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
pytest==8.3.2
pytest-django==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
webencodings==0.5.1
